refactor: extract dependency rules into analyzer/rules.rs (#190)
Moves RuleViolation + check_dependency_rules out of analyzer/mod.rs.
